Use ContinuationCore
