+++ Source/AwsCommonRuntimeKit/io/retryer/CRTAWSRetryStrategy.swift
@@ -25,16 +25,33 @@ public class CRTAWSRetryStrategy {
     }
 
     public func acquireToken(timeout: UInt64 = 0, partitionId: String) async throws -> CRTAWSRetryToken {
-        return try await withCheckedThrowingContinuation { (continuation: AcquireTokenContinuation) in
-            let crtRetryStrategyCore = CRTRetryStrategyCore(continuation: continuation)
-            crtRetryStrategyCore.retainedAcquireTokenFromCRT(timeout: timeout, partitionId: partitionId, crtAWSRetryStrategy: self)
+        return try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<CRTAWSRetryToken, Error>) in
+            let continuationCore = ContinuationCore(continuation: continuation)
+            let retainedContinuation = continuationCore.passRetained()
+            if (partitionId.withByteCursorPointer { partitionIdCursorPointer in
+                aws_retry_strategy_acquire_retry_token(rawValue,
+                        partitionIdCursorPointer,
+                        onRetryTokenAcquired,
+                        retainedContinuation,
+                        timeout)
+            }) != AWS_OP_SUCCESS {
+                continuationCore.release()
+                continuation.resume(throwing: CommonRunTimeError.crtError(.makeFromLastError()))
+            }
         }
     }
 
     public func scheduleRetry(token: CRTAWSRetryToken, errorType: CRTRetryError) async throws {
-        try await withCheckedThrowingContinuation({ (continuation: ScheduleRetryContinuation) in
-            let crtRetryStrategyCore = CRTRetryStrategyCore(continuation: continuation)
-            crtRetryStrategyCore.retainedScheduleRetryToCRT(token: token, errorType: errorType)
+        try await withCheckedThrowingContinuation({ (continuation: CheckedContinuation<(), Error>) in
+            let continuationCore = ContinuationCore(continuation: continuation)
+            let retainedContinuation = continuationCore.passRetained()
+            if aws_retry_strategy_schedule_retry(token.rawValue,
+                    errorType.rawValue,
+                    onRetryReady,
+                    retainedContinuation) != AWS_OP_SUCCESS {
+                continuationCore.release()
+                continuation.resume(throwing: CommonRunTimeError.crtError(.makeFromLastError()))
+            }
         })
     }
 
@@ -47,4 +64,31 @@ public class CRTAWSRetryStrategy {
     deinit {
         aws_retry_strategy_release(rawValue)
     }
+}
+
+private func onRetryReady(token: UnsafeMutablePointer<aws_retry_token>?,
+                          errorCode: Int32,
+                          userData: UnsafeMutableRawPointer!) {
+    let crtRetryStrategyCore = Unmanaged<ContinuationCore<()>>.fromOpaque(userData).takeRetainedValue()
+    if errorCode != AWS_OP_SUCCESS {
+        crtRetryStrategyCore.continuation.resume(throwing: CommonRunTimeError.crtError(CRTError(code: errorCode)))
+        return
+    }
+
+    // Success
+    crtRetryStrategyCore.continuation.resume()
+}
+
+private func onRetryTokenAcquired(retry_strategy: UnsafeMutablePointer<aws_retry_strategy>?,
+                                  errorCode: Int32,
+                                  token: UnsafeMutablePointer<aws_retry_token>?,
+                                  userData: UnsafeMutableRawPointer!) {
+    let crtRetryStrategyCore = Unmanaged<ContinuationCore<CRTAWSRetryToken>>.fromOpaque(userData).takeRetainedValue()
+    if errorCode != AWS_OP_SUCCESS {
+        crtRetryStrategyCore.continuation.resume(throwing: CommonRunTimeError.crtError(CRTError(code: errorCode)))
+        return
+    }
+
+    // Success
+    crtRetryStrategyCore.continuation.resume(returning: CRTAWSRetryToken(rawValue: token!))
 }
